cleanup projection tests
