+++ tests/Projection/MySQLEventStoreProjectionTest.php
@@ -12,216 +12,33 @@ declare(strict_types=1);
 
 namespace ProophTest\EventStore\Projection;
 
-use ArrayIterator;
-use Prooph\Common\Messaging\Message;
-use Prooph\EventStore\Exception\RuntimeException;
-use Prooph\EventStore\Exception\StreamNotFound;
-use Prooph\EventStore\StreamName;
-use ProophTest\EventStore\Mock\UserCreated;
-use ProophTest\EventStore\Mock\UsernameChanged;
-use ProophTest\EventStore\PDO\Projection\AbstractMySQLEventStoreProjectionTest;
+use Prooph\Common\Messaging\FQCNMessageFactory;
+use Prooph\Common\Messaging\NoOpMessageConverter;
+use Prooph\EventStore\PDO\MySQLEventStore;
+use Prooph\EventStore\PDO\PersistenceStrategy\MySQLSimpleStreamStrategy;
+use ProophTest\EventStore\PDO\Projection\PDOEventStoreProjectionTestCase;
+use ProophTest\EventStore\PDO\TestUtil;
 
 /**
  * @group pdo_mysql
  */
-class MySQLEventStoreProjectionTest extends AbstractMySQLEventStoreProjectionTest
+class MySQLEventStoreProjectionTest extends PDOEventStoreProjectionTestCase
 {
-    /**
-     * @test
-     */
-    public function it_links_to_and_loads_and_continues_again(): void
+    protected function setUp(): void
     {
-        $this->prepareEventStream('user-123');
-
-        $projection = $this->eventStore->createProjection('test_projection');
-
-        $projection
-            ->fromStream('user-123')
-            ->whenAny(
-                function (array $state, Message $event): array {
-                    $this->linkTo('foo', $event);
-
-                    if ($event->metadata()['_aggregate_version'] === 50) {
-                        $this->stop();
-                    }
-
-                    return $state;
-                }
-            )
-            ->run();
-
-        $streams = $this->eventStore->load(new StreamName('foo'));
-        $events = $streams->streamEvents();
-
-        $this->assertCount(50, $events);
-
-        $events = [];
-        for ($i = 51; $i < 100; $i++) {
-            $events[] = UsernameChanged::with([
-                'name' => uniqid('name_'),
-            ], $i);
+        if (TestUtil::getDatabaseVendor() !== 'pdo_mysql') {
+            throw new \RuntimeException('Invalid database vendor');
         }
-        $events[] = UsernameChanged::with([
-            'name' => 'Oliver',
-        ], 100);
 
-        $this->eventStore->appendTo(new StreamName('user-123'), new ArrayIterator($events));
+        $this->connection = TestUtil::getConnection();
+        $this->connection->exec(file_get_contents(__DIR__.'/../../scripts/mysql/01_event_streams_table.sql'));
+        $this->connection->exec(file_get_contents(__DIR__.'/../../scripts/mysql/02_projections_table.sql'));
 
-        $projection = $this->eventStore->createProjection('test_projection');
-
-        $projection
-            ->fromStream('user-123')
-            ->whenAny(
-                function (array $state, Message $event): array {
-                    $this->linkTo('foo', $event);
-
-                    if ($event->metadata()['_aggregate_version'] === 100) {
-                        $this->stop();
-                    }
-
-                    return $state;
-                }
-            )
-            ->run();
-
-        $streams = $this->eventStore->load(new StreamName('foo'));
-        $events = $streams->streamEvents();
-
-        $this->assertCount(100, $events);
-    }
-
-    /**
-     * @test
-     */
-    public function it_emits_events_and_resets(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $projection = $this->eventStore->createProjection('test_projection');
-
-        $projection
-            ->fromStream('user-123')
-            ->when([
-                UserCreated::class => function (array $state, UserCreated $event): void {
-                    $this->emit($event);
-                    $this->stop();
-                },
-            ])
-            ->run();
-
-        $streams = $this->eventStore->load(new StreamName('test_projection'));
-        $events = $streams->streamEvents();
-
-        $this->assertCount(1, $events);
-        $this->assertEquals('Alex', $events->current()->payload()['name']);
-
-        $projection->reset();
-        $this->assertEquals('test_projection', $projection->getName());
-
-        $this->expectException(StreamNotFound::class);
-        $this->eventStore->load(new StreamName('test_projection'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_emits_events_and_deletes(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $projection = $this->eventStore->createProjection('test_projection');
-
-        $projection
-            ->fromStream('user-123')
-            ->when([
-                UserCreated::class => function (array $state, UserCreated $event): array {
-                    $this->emit($event);
-                    $this->stop();
-
-                    return $state;
-                },
-            ])
-            ->run();
-
-        $streams = $this->eventStore->load(new StreamName('test_projection'));
-        $events = $streams->streamEvents();
-
-        $this->assertCount(1, $events);
-        $this->assertEquals('Alex', $events->current()->payload()['name']);
-
-        $projection->delete(true);
-
-        $this->expectException(StreamNotFound::class);
-        $this->eventStore->load(new StreamName('test_projection'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_ignores_error_on_delete_of_not_created_stream_projections(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $projection = $this->eventStore->createProjection('test_projection');
-
-        $projection
-            ->fromStream('user-123')
-            ->when([
-                UserCreated::class => function (array $state, UserCreated $event): array {
-                    $this->stop();
-
-                    return $state;
-                },
-            ])
-            ->run();
-
-        $projection->delete(true);
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_on_run_when_nothing_configured(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $projection = $this->eventStore->createProjection('test_projection');
-
-        $projection->run();
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_trying_to_run_two_projections_at_the_same_time(): void
-    {
-        $this->expectException(\Prooph\EventStore\PDO\Exception\RuntimeException::class);
-        $this->expectExceptionMessage('Another projection process is already running');
-
-        $this->prepareEventStream('user-123');
-
-        $projection = $this->eventStore->createProjection('test_projection');
-
-        $eventStore = $this->eventStore;
-
-        $projection
-            ->fromStream('user-123')
-            ->whenAny(
-                function (array $state, Message $event) use ($eventStore): array {
-                    $projection = $eventStore->createProjection('test_projection');
-
-                    $projection
-                        ->fromStream('user-123')
-                        ->whenAny(
-                            function (array $state, Message $event): array {
-                                $this->linkTo('foo', $event);
-
-                                return $state;
-                            }
-                        )
-                        ->run();
-                }
-            )
-            ->run();
+        $this->eventStore = new MySQLEventStore(
+            new FQCNMessageFactory(),
+            new NoOpMessageConverter(),
+            TestUtil::getConnection(),
+            new MySQLSimpleStreamStrategy()
+        );
     }
 }
